Splitting congrollers, Adding delete functionality for users in admin area

+++ src/main/java/com/poisonedyouth/springSecurity/service/RoleService.java
@@ -17,7 +17,7 @@ public class RoleService {
 		return roleRepository.findAll();
 	}
 
-	public Role findRoleById(int id) {
+	public Role findRoleById(Long id) {
 		return roleRepository.findById(id).get();
 	}
 }
